slamd-96: Fixed apriori and categorical tests and moved them to appropriate file
